+++ console/src/components/Common/NotSupportedNote/index.tsx
@@ -1,12 +1,5 @@
 import React from 'react';
-import { Driver, currentDriver } from '../../../dataSources';
-
-const driverToLabel: Record<Driver, string> = {
-  mysql: 'MySQL',
-  postgres: 'PostgreSQL',
-  mssql: 'MS Server',
-  bigquery: 'Big Query',
-};
+import { Driver, currentDriver, driverToLabel } from '../../../dataSources';
 
 type NotSupportedNoteProps = {
   unsupported: Driver[];